Add support for parsing assertions in ad-hoc goldens
So far, a select and assert will be very similar. I need to be able to
use "where" for assertions in some cases anyway, and I also need select
to be able to select the ids of created entities, so that all fits very
naturally in the existing query structure.