Extract golden segmentation into module
I want to be able to run the parser script on the goldens, but the
goldens contain expected output in addition to the queries. The runner
did split them, but the parser script did not. By extracting this into
a module, it can be reused, and the parser script can parse multiple
queries in one run too.

This does not yet update the runner to also use the segmentation module.