Define parser for binary transaction requests
Finding a good name for this thing is hard. It is not a transaction; the
transaction is what gets created as a result. It is not a query either,
because queries only read things, but this is a write. It does include a
query as a subset though.

"Assertion" is close, but the thing does include a query part, so
assertion in itself is not the full story.

Perhaps the difficulty to name this thing is a sign that it's the wrong
thing. Perhaps I should not merge queries into writes after all, and let
the client library deal with it. Then a transaction request would really
just be a list of statements.

But for now, let's see where this goes.