Extend operation support in "execute" binary
Now we can do either reads (type 0, "query::Query"), or writes (type 1,
"query::QueryMut"). Writes are not yet implemented in this commit, only
the logic to detect that we are dealing with a write.

The parser/serializer also needs to be updated, because the binary
format now has a prefix to indicate the query type.