Create temporaries for large values in parser
When the query contains large values, we need to put values on the stack
pool (the pool of temporary values) to be able to construct a Value from
them that can be used in queries.

For assertions later, those temporary values need to be converted into
persistent values, on the underlying pool. But for now, we put all
values into the stack pool. That may be better anyway, because even
though a value occurs in an assertion, that does not mean it will be
persisted, for instance if the query returned zero results.