Explain why Noblit does not have a VAET index
When watching some more Datomic talks, it struck me that the design is
very much oriented towards dynamic typing. That is no surprise, given
that Clojure is also dynamically typed. My personal preference for the
amount of static typing is probably to have stronger types than what is
currently mainstream (although there is a second wave of gradually typed
languages now, following the rise of dynamically typed languages). Using
type-level-natural-indexed vectors is rarely worth the effort, but e.g.
creating types and parsing json dicts into them is very worthwhile.
Clojure seems to go more of the Python way where everything is a dict
with string keys and "any" values. But for Noblit I'd like to be able to
map query results directly into strong types, without having to parse
them, and without the opportunity for a parse failure due to a mismatch
between the shape of the data and what the application expects.