Insert datoms after producing them in tranaction
This currently fails, because the values are still referring to
temporary values on the stack pool, but in order to persist the datoms,
we need to persist the large values too.

Fortunately, there is an assertion that catches this, and we don't get
silent data corruption. Next up: add a phase in between that translates
the stack pool values (and the datoms containing them).