Make golden runner interactive
We need to run every query (read or write) against the same process, in
order for the assertions to have any effect on the subsequent read. For
that, we keep one process open, feed it a query, read its response, etc.

Currently the executor does not stream stdin, so the runner now hangs.