Fix expected output and query in define_attributes
Assertions run once for every output of the "where" part. That means
that you alway need to query something to get the assert to work, and
you need to ensure the output has cardinality 1 when you want to do a
single insert.

I should special-case the case of an empty where-clause to just run the
insertions once. On the one hand, I don't like special cases; this one
breaks symmetry and regularity. On the other hand, this case is entirely
reasonable.