Allow exfiltrating temporaries from stack pool
The stack pool borrows the pool immutably. But to persist the
temporaries, I need a mutable pool. But to make the pool mutable, I
need to drop the stack pool. The solution is to drop the stack pool
but rescue the temporaries. We no longer need to resolve arbitrary
values, only the temporaries. Then they can be persisted.