Add some examples from bug tracker as golden
There are still some things in there which are not supported, but the
majority already runs!

I should write the goldens as literate examples, that would be even
better than the current form. Then they can double as examples with
explanations, while at the same time verifying that the example behaves
as it should, and while doubling as a test.