Begin reworking temporaries and the temp heap
I just realized: for the temp heap, there is no reason that we should
borrow the underlying heap before we insert the entries. Also, the idea
of stacking that I had turns out to not be useful. There is only one
point at which temporaries come into scope and go out of scope. The
solution then, is to split the concerns:

 * A collection of constants, that we can build during query parsing,
   and use during datom insertion.

 * A temp heap, that combines an underlying heap with a set of
   temporaries, to provide uniform lookup over the two.

As a first step towards that, change the TempHeap and Temporaries
definitions to support this. This leaves the calling code broken, fixing
that is the next challenge.