Work more towards splitting temporaries and heap
The library typechecks now, but the applications need to be updated. As
part of this, it turns out there was one place where I used the push/pop
behavior of the temporaries stack: to resolve named attributes. But for
that, we might as well put the name on the temporaries stack during
query parsing, and then use the normal logic. So that is what I did now.