Fix usage of temporaries in bin/main
Not yet in the test executor, but at least the dummy main program works
again now.

As a side effect, lift the &mut restriction on queries resolving named
attributes to by-id attributes. Now that attribute names are already
temporaries before starting the query, there is no need to manipulate
the temp heap in the view to add the attribute name for its lookup.
Therefore, this can work with a read-only view, which aligns much better
with the high-level mutability idea: that a query should not require
&mut access.