Remove db.transaction.time builtin attribute
I am no longer convinced that this is useful. A timestamp (as in UTC
time) may be useful, but it is not up to me to decide whether it should
be included; applications may want to omit it for reproducibility
reasons.

What could be useful is not the timestamp, but a sequence number; a
transaction counter. ("Time" is a bit of a misnomer.) This is still
reproducible, it could be added by the database without going down the
time rabbit hole.

But then again, the transaction id is already a monotonically increasing
counter. Only it may skip values, so you can't do +1 or -1 on it and
expect to get another transaction id. In fact, as transaction ids are
even, you would need +2 or -2, but even that will not always work. But I
am not convinced that having a counter that does satisfy that property
is worth the cost of an extra datom in every transaction, or the
inherent complexity of having more builtins. So let's remove it.