Add module for generating fresh ids
The two separate counters with +2 logic work, but the logic has to be
duplicated. It is too simple to justify full machinery like I am adding
in this commit, but repeating it everywhere is not nice either.

But apart from that, I would like to consume the id space evenly, and
not have them be completely separate for entities and transactions.
Although that may be useful, it also limits the space of usable ids
(although stealing one bit is probably fine). But more importantly,
generating even ids ensures that if you want to add data to
transactions, those datoms get sorted close to other new datoms in the
eavt and aevt indexes, which makes htree insertion a bit more efficient,
because there is a larger consecutive chunk.

So there is an slighly more complicated way to hand out ids evenly,
which I am now adding in a new module with tests. By encapsulating this,
it can later also be added easily as part of a database header.