Make commits explicit with new Head struct
A database is a collection of trees. The trees are persistent (in the
sharing-structure data structure sense), so nodes are immutable and we
only append new ones, but we need to track what the current roots are
*somewhere*. In addition, there are the counters to generate fresh ids
for entities and transactions. Those two (the roots and counters) form
the mutable state of the database. Now we can make that explicit with a
"Head" struct. New data gets added, the head moves.

It would be nice to encode the steps to perform a transaction a bit
better in the type system, and to make it less error-prone. I am
confident something is possible, but for now, let's do this
encapsulation of id generation first.