Fix ids in test expectations after idgen change
I changed the way ids are allocated, but I forgot to also update the
test expectations for the new algorithm. Ids are now closer together,
so it makes sense that the valuse are lower. The entity ids are mostly
sequential, but sometimes they skip an even to accomodate the
transaction id.