Add struct to manage transactions
Instead of juggling around tree roots, we can pack the things together
in a struct. This makes it a bit easier to create transactions. Next
steps would be to make it track the collection of new datoms too. Then
the transaction id would not need to be exposed, nor would the
collection.

As a final step, it would be nice to only allow a single transaction to
exist at a time, to ensure that you can't commit them in different
orders that they were created.