Fix bug in constructing Value from u64
The size check was wrong, it should have checked that

    val <= ~Value::TAG_MASK

where ~ is bitwise negation. The mask bits are the most significant
bits, so the greatest value still allowed is the inverted mask. Checking
that

    val < Value::TAG_MASK

is wrong, because having the least significant mask bit set also passes
that test.

Rather than writing down the rather cryptic comparison, I opted for
checking that the mask bits are zero. That is more obvious to me, and it
is also what the assertion checks anyway.