Bring "files" documentation a bit more up to date
I wrote that page a long time ago, when the implementation was not as
advanced as it is now. At that point I still wanted to have a journal.
But there is no need for a journal, we can just update the indexes at
every transaction, that is a lot simpler, because all datoms are always
indexed, and because we need to write indexes at some point anyway.

Also, because the indexes are persistent data structures, we don't need
a journal to ensure atomic updates.

But what we do need instead is the "head", that contains the page ids of
the root nodes of the index trees. If we can do an atomic update of the
head, then we get atomic transactions for free.