Write down idea for aggregations
For a database to be useful, grouping and aggregations are quite a basic
need. I spent some time thinking about how it would fit into Noblit's
data model, and fortunately it falls out naturally, if results are
ordered. There are things we can build on top of ordered results:

 * Returning only distinct results.
 * Grouping and aggregations over the items in a group.

The former is really just a special case of the latter, with no
variables aggreated over and all variables grouped by. For now let's
do nothing special to implement distinct results. The real important
thing is to implement "order by", because grouping sorted data is
trivial.