Write down idea for new query plan structure
My first attempt at doing query plans was rather ad hoc. Not only the
plan, but also the planner. Since then, I've come to realize (credit to
the CMU online database course for this) that the way to plan queries,
at least in theory, is to generate all plans, have a way to estimate
their cost, and then to take the least costly one. This is nice, because
you can start with simple heuristics for cost, and improve them over
time (possibly also taking statistics about the data into account).

Of course generating all possible query plans is prohibitive. But,
generating a plan can be done incrementally, where we have to make a
decision at every step. And as we progress in the plan, our estimate of
its cost gets more accurate (or rather, complete). This smells very much
like a tree search problem, and it can probably be solved like that,
with some heuristics to explore the promising parts of the tree first.

If I want to add "order by" support later, that is also very natural to
do with such a query plan. Because variables come from scans over sorted
indexes, the results are ordered by the order of the scans. We can then
test whether the result order of a plan is compatible with the desired
sort order. If it is not, it is not a valid plan for the ordered query.
So adding "order by" to the query just prunes the tree of plans.

Apart from this, I was just toying around with the loop structure again.
It is possible to have loops that define two variables, not just one; a
full index scan can write both variables. So then I started writing
things down, and making scans (loops) the fundamental part, decoupled
from variables, seemed reasonable. A scan could either read one variable
(to constrain the range) and write another, or it could write both. But
then I still needed some kind of filter construct, besides the loops.
Then I realized that there is a third kind of data flow that would fit
here: a scan that reads two variables and writes none. It is essentially
an existence test; it iterates once for every datom that exists, so if
the datom does not exist, you get zero iterations, and the inner yield
does not execute. So with these three combinations (out-out, in-out,
in-in), everything falls out naturally!

This leaves one to wonder, would out-in be a sensible fourth scan? But
no, the sort order matters. Inputs go on the left first, so if we have
a single input, it should be the leftmost one of the index.