Add invariant check for new query plan
Evaluating query plans is going to mutate some buffer of slots in place.
Even if that works out to be initialized in Rust, logically these slots
have a notion of being initialized or not, and furthermore, we should
allow only a single scan to write to a slot. These things are not
checked by Rust, but we can manually write a check for this. We can run
it in debug mode, and it is useful to have for tests and for fuzzing,
although I do not yet have any of these.