Add new planner skeleton, drop types from plan
It turns out, to plan the query, we do not need any knowledge about the
types at all. Of course, for the unification of the variable in
different places to make sense, the types must match. But at runtime, we
can always compare every Value, even if two values have different types.

It turns out that the query planner does not need access to the types in
order to make then plan. Maybe for the cost estimation it would come in
handy (integers are likely inline, strings likely external), but let's
not worry about that for now.