Make planner generate an initial plan
For a given ordering of the where statements, there is not that much
freedom left. The ordering of the statements determines which slots are
initialized at which point. And whether zero, one, or two of the
referenced slots are initialized when we get to a statement, determines
which indexes we can use to fill the slots. Sometimes there is only one
option, sometimes there are mutiple. But the referenced slots or the
flow (how many slots are inputs and how many are outputs) do not change.
So what we can do is generate one single plan, and then enumerate all
alternatives by lexicographically "incrementing" every scan.

As a first step towards this, generate *a* plan. Then next we can look
into enumerating alternatives.

Finally, it would also be possible to combine this enumeration with
permuting the statements to exhaustively generate all possible plans.