Store (entity, value) in Scan, not (first, second)
Depending on the index, the sort order of entity and value is different.
But it is convenient to be able to call them entity and value anyway,
because then we can cycle through scans while keeping the entity and
value slots in the scan fixed. If we store first and second, we would
have to swap them if we change aevt into avet when cycling.

For in/out it *is* useful to know which of the two slots is the input
and which is the output, but we can compute that on the fly.