Implement debug printing for new query plans
I've been writing a lot of code, but I can't actually *see* what the
query plans look like, and I haven't written the new evaluator yet, so I
can't run them either. To get some intermediate feedback, being able to
print plans is very useful.

This also adds prints for the queries in the main program, which is a
sandbox anyway, and it looks like the queries get planned correctly! At
least, there is some valid plan for that query.

The next step is to generate *all* plans for a query and rank them.