Prefix golden test expectations with "> "
Segmenting on box drawing characters was neat, but if I want to test
query plans in the goldens, then either I need to put a line around the
expected plan, or I just switch to a prefix for expectations. The latter
seems better in the long run, as it always works.

There is still some box-drawing magic to recognize that the output of
the execute program is done. I should probably have a terminator
indicator there too, but we can do that at a later time.