Add "explain" support to executor and golden tests
To make sense of the query plans generated by the new query planner, I
want to have some way to submit a query, and get the plan or plans. In
SQL, this is traditionally done with an "explain" query, and we can copy
that here, it maps well to Noblit's query model too.

To plan a query though, we need the schema, so we need to set up an
interesting database. And the easiert way to do that right now, is
though the Golden tests. Also, extending the "execute" binary with
support for "explain" is easy. So I opted to add this to the golden
tests, not so much for testing the planner (although that is a nice side
effect), but mainly so that I can look at the plans.

The output format is not yet ideal, for example constants are printed
using the Debug impl for Value, which is not yet very user-friendly. But
we can change that in the future and update the goldens.

Also, the plans are expecetd to change with improvements to the planner.
We can update the goldens there too. At least we should be able to see
at a glance if the plan is an improvement or not.