Write down idea for explicit query optimizer
Exploring all permutations of statements in the optimizer is going to
be tough, and to select which one to use, we would need some good
heuristics to estimate the performance of a plan. That would probably
require statistics to make a good estimate. And as there are too many
permutations for most queries, we would need some kind of
heuristics-driven tree search. All of this sounds pretty complex.

So what if we just don't bother with query optimization. The current
planner is very straightforward. I still need to explore which index to
use, but for a given permutation of statements, we will do the same
kind of scans. This is simple, which has two advantages:

 * It is fast. There is no trade-off between spending time finding a
   better plan, and spending time evaluating that plan.

 * It gives users a lot of control, and predictible performance. The
   query specifies mostly what the plan will look like, and if you want
   a different plan, you can change the query.

So if we drop an advanced optimizing planner, but instead go with the
control that a straighforward planner gives us, suddenly new things
become possible:

 * If we would know which plan was best, we actually have a way to tell
   the database to use that plan.

 * To find which plan is best, we can try different plans, and measure.

 * If we are not constrained by a deadline to answer a query quickly, we
   can afford to spend some time optimizing the query.

 * Exploring alternatives to a query is very simple for Noblit's
   queries: the only thing that matters is the order of statements, so
   we only need to explore permutations of those.

Exploring *all* permutations is still going to be infeasible quickly,
especially if we want to profile each plan. So probably we still need a
way to estimate performance, so we can rule out bad candidates and only
profile promising ones. And perhaps even a way of doing guided search.
But if you do profiling too, that process could be profile-guided.
Either way, we can start with something basic (generate all
permutations, see which one is fastest), and work towards a more
advanced query optimizer later.