Add end-of-response marker to golden executor
Guessing based on box drawing characters, or "yield", is a bit fragile.
I want to print multiple plans, but if "yield" is the end marker, then
that does not work. So use an explicit end marker. I could have used
some magic text, but ascii does have an archaic byte for it, so why not
use that? Because it is not printable, would be a very good reason, but
when running under the test runner we can filter it.