Move IdGen into head module
It is one of the parts of the head, it neatly fits there, and having its
internals there makes it easier so serialize and deserialize in a single
place. The "idgen" module was a bit weird on its own anyway, like
kitchen sink "utils" module with only one thing in it. It feels more at
home with the other parts of the head.