Add support for serializing the head
Working with fixed-size slices in Rust is surprisingly tedious. This
stuff should be what Rust is good at, but alas, we need to
inconveniently slice all the things and then pray to the LLVM gods to
optimize all of that away again, even though everything could typecheck
nicely if indexing with constants would produce a fixed-size slice.

I haven't checked if LLVM does actually optimize these things. If it
turns out to be a bottleneck at some point, we can optimize it then.