Add modue for writing the on-disk format
Up to now, Noblit has been an in-memory database, simply because
persistence had not been implemented yet. But now for tests, it is
useful to be able to generate databases somewhere and then use them,
I don't want to have to populate them in the test program all the time.
So we need to start doing persistence.

I am not going to do full persistence, with a file-backed store and heap
yet. What we can do instead, is have a "packed" format that dumps the
in-memory database to a file, and we can read the file back into memory
later. It should be suitable for later reading pages from it on-demand,
without having to read it into memory, but for now that is fine.

Also, I can already make the format suitable for writing the mutable
database with later. The format will be very similar, only the mutable
one will be split in three files (store, heap, head). The format of the
store and heap is already fixed, the format of the head could be almost
the same for both packed and unpacked formats.