Write head and store in packed on-disk format
Now we almost have enough to reconstruct the entire database. Only we
still need the heap. For that, I need a way to walk the heap values. Or
maybe I'll just add a method to dump it to an io::Write. The same might
work for the store actually. That would be nicer than relying on all
these page ids.