Add skeleton for reading database from file
This is not as efficient as it could be, we are doing double buffering
here (or triple buffering even, if the input is a BufReader). But that
is fine for now, performance is not yet a concern, and this is one-time
IO to load a database into memory. For now, I just want to be able to
save and load things.

Also, panic on format error, instead of reporting a proper error, such
that I can use io::Result as the return type. I should add a custom
error type to Noblit that also includes format errors, but that can be
added later. For now I just wan to be able to dump and write a database,
adversarial inputs are not a concern at this point.