Add alignment assertion to transmute_slice
There was a to do to check this for longer, because indeed the function
is very unsafe. The source of the pages should ensure that they are
aligned properly. For example, when memory mapping backing files, they
should be mapped at the right offset.

But one of my current store implementations is a Vec<u8>. There it is
more difficult to guarantee alignment. I once argued that allocators
should support over-alignment (at the time for performance reasons), but
the proposal was rejected. So it is impossible to reliably get a Vec<u8>
with e.g. 8-byte alignment. But maybe we don't need Vec<u8>, and we can
use Vec<u64> as the backing store. Then we do get the right alignment.

Alignment has not been a problem in practice yet. Perhaps because x86
will usually happily do unaligned loads, or perhaps because the
allocator gives us 8-byte aligned buffers anyway. I don't care about
non-x86 architectures, so maybe this is not a problem in practice. But
still, it would be nice to fix.

I am now trying to run the tests and executables under Miri, and indeed
Miri does detect the wrong alignment and complains about it. After
adding the assertion, it hits the assertion under Miri instead. I think
it can be fixed with Vec<u64>.

I also realize now that the 563-byte page size that is used in tests and
fuzzing is incompatible with 8-byte alignment that u64 requires. 563 was
nice *because it catches implicit alignment dependencies* by being
prime, but alas, we will have to drop it. 568 would work, which at least
is a prime multiple of 8, which is the next best thing we can do if we
need a multiple of 8 anyway. But that is for a later commit.