Change page size 563 to 568 for 8-byte alignment
We cannot have all of the following:

 * The page size is a prime.

 * Pack every page densely in the store without padding.

 * Ensure 64-bit integers in the page (in datoms, and the page ids) are
   aligned to 8 bytes in every page.

I don't want to drop the second requirement (doing unaligned reads
everywhere would be a real pain). The second requirement is not a hard
requirement, but it would be somewhat inconvenient to give up, because
then we have to insert padding manually everywhere. Therefore, let's
give up the first requirement, and change the page size to be a multiple
of 8. Having a prime page size was nice *because* it often leads to
nonzero residues modulo powers of two, which exposes bugs where
alignment is implicitly assumed but not explicitly checked. By moving to
a multiple of 8, we need to give some of this up; the residue modulo 8
will always be 0. But modulo greater powers of two it will not always
be, so we still have that.