Add file loading capability to "execute" binary
This allows using a database from a file as a starting point, instead of
the default empty database. This means that we no longer have to insert
all the data to query from on the fly.

I added a Noblit export function to Mindec, my music metadata indexer,
so now I have a 13MB database with ~15000 entities, which should be a
nice test bed for testing queries, but especially for running benchmarks
for the new query planner and evaluator, on some real-world data.

It also opens up the possibility for running golden tests against known
databases that would be difficult to set up, and to include databases as
regression tests if necessary.