Add new query plan evaluator
This is partially copied from the existing evaluator in query_plan,
but adapted to the new query plans. Many things become simpler, and
more elegant this way. One key insight is that there are only four
cases relevant for evaluation: for the entity and value, whether it is
an input or output. While there are nine combinations of flow and index,
these map to the four entity-value flows. By making the evaluator deal
only with those four, things become quite manageable.

Having unified slots that store Values, instead of the ad-hoc
"definitions", also makes things a lot nicer. The pieces are more
orthogonal, not as ad-hoc. Also, the new plans do not confuse scans
with slots any more. And constants are no longer a special case either,
as they get stored in a slot.

The one thing that does still bother me is that we need to construct
the initial iterator, and that iters[0] is a special case. It could be
avoided by having one more iterator that produces a single value at
index -1. Only index -1 is impossible, so we'd need to swap the order.
I'm not sure if all of this is worth it in the end. For now I suppose
this works fine, the special case is not terrible either.

This does not yet implement the iterator. I will copy over that one
next, and handle selections.