Add iter for new evaluator, use for one query
Some things become a bit more involved -- we need to run type inference
manually. But on the other hand, types are erased, so that we can run it
as a separate pass and then forget about it is also quite neat. Also,
add a few helper methods to make printing the result table easier.

This uses the new evaluator for the last query in the "main" binary.
And it works, yay! The order of the results is different, but the plan
is also different, so that was to be expected.

Next up is using this everywhere, and perhaps making it a bit more
ergonomic. After that, the old plans and evaluator can be deleted.