Make golden test diff output more useful
If it prints a huge block, I can't tell where in the block it went
wrong. Just print added/removed instead. There is no fancy diffing,
so if things run out of sync, it never recovers. But it is possible
to fix the desired outcome one line at a time now.