Prevent Miri build from invalidating CI cache
Caching did speed up the build, but because Miri produces new files all
the time, it also uploads the full cache at every build, which takes
about a minute. That wastes the caching win. I hope that by deleting
the files that change all the time, the cache will not change, as it
contains only deps, and the Noblit build is fast anyway.