Optimize only indexes for now, not statements
The impact of which index to use is by far not as much as that of the
order of the statements, but for a good statement order, all of my test
queries take on the order of 10 microseconds to execute, and there I do
see something like a 50% difference between the p10 and p90 execution
time. So even if the difference is in the order of microseconds, that
can be a big relative difference for fast queries.