Add explore-exploit based index plan optimizer
The idea is to profile every option, but we should spend more effort on
promising candidates. One of the ways to do this is to take the bounds
of a confidence interval as the expected reward, instead of taking the
center of the confidence interval. As we gather more data, the interval
shrinks, so in every iteration, we learn something valuable: either we
exploit the good plan, and become more confident that it is good, or we
explore a bad plan, and that shrinks its bounds so we will waste less
time on it in the future.

That's the idea, at least. But then this is complicated by the fact that
measuring performance is extremely difficult, and it is affected by all
kinds of influences that cannot be modelled as gaussian noise but that
nonetheless distort the results. I went with taking the fifth percentile
after all (a low percentile because we care about the best case,
everything on top of the best case is noise -- but not the minimum,
because extreme percentiles have high variance), but there is definitely
room for improvement. I don't estimate a confidence interval at all, but
just scale the bounds with 1/sqrt(n) (which does coincide with the
scaling factor in the width confidence interval of the sample mean), and
that seems to work wel. 1/sqrt(sqrt(n)) leads to more unifor exploration
(more explore, less exploit).