Add profiling query macro-optimizer
This adds the first steps towards the profiling query optimizer that
optimizes statement order. The index optimizer was just toying around,
it does not appear to matter too much and I don't know how to integrate
it either way. But the statement order can make a huge difference.

I started with a basic way of exploring the tree of possibilities, which
appears to work well in practice: explore first where the duration of
the query up to that point is the least. This assumes that adding a
statement at the end can only make queries slower, not faster. I think
this is a reasonable assumption, especially when the indexes are fixed.
When you take indexes into account, there can be non-local effects (e.g.
adding a second statement switches an earlier scan from AEVT to EAVT and
for some reason that speeds up the first scan too). But still, I think
it is a reasonable assumption, and it practice it appears to work.

Another way to look at this, is that the tree of partial query plans
(where the leaves are the full, final query plans) is a graph, and -- if
the above assumption about time growing monotonically with number of
statements, is correct -- the edges are labeled with the extra time that
that adding the statement incurs. Then we do Dijkstra on the graph to
find the closest leaf.

I was worried initially that the cost of early nodes would be so cheap
that this algorithm would spend a lot of time exploring near the root,
going wide instead of deep, before it finds anything. But that turned
out not to be the case in practice, it does find a leaf really quickly.

The only issue is that sometimes it does go and explore what I think is
a double (or maybe triple?) full index scan with no shared variables.
And then suddenly the execution time shoots up from microseconds to
dozens of seconds, and it gets stuck. I should add some shortcut to
detect those cases early and bail out.