Bail out on slow options in query macro optimizer
This alleviates the problem where we happen to try a candidate which
suddenly generates a cartesian product of two full indexes and then we
are stuck on it for a long time, even though it is obvious that it is a
bad option. Now, if the time is a lot worse than then know best
solution, we stop exploring, there is little point in it then.

In fact, now I think about it, even if it is fast enough, we may want to
skip it, because it is not the fastest, so if the monotonicit assumption
holds, then anything in this branch of the tree is going to be worse ...
I think. But thinking about this more, I'm not so sure. It may be that
there is a "bad" statement that as soon as you include it blows the
times up. And depending on how the statements work, it may be
advantageous to include it sooner rather than later. Hmm... Maybe I
should also not bail out entirely but just reduce the number of
iterations.

Anyway my mind is no clear enough to think deeply about it at this
point. For now this works nicely. For every query I enter the results
look plausible. The output is not always the same, but it only differs
in orders that you'd expect to be irrelevant (e.g. different attributes
of the same entity). So overall I think this is a good way forward.