Prioritize query nodes by duration lower bound
The previous method of measuring all iterations for a candidate was a
bit hacky, especially when you explore a bad candidate too often. My
approach of tracking fast ones and bailing out early was a bit of a
hack. Estimating a lower bound is a much more natural solution that gets
us the same behavior: good candidates get explored more, and if we do
add a bad candidate, we only explore it a few times until it is clear
that it is bad. But if we had a spike in an initial measurement and it
is actually good, we will also find out quickly and explore it further.