Merge micro-optimization and macro-optimization
The MinDuration exploration scheme for macro-optimization was working
well. We can use it for micro-optimization too. While exploring, we
don't only track the partial query (the included statements, and their
order), but also all of the possible plans for that query. The best time
for a partial plan is the best time of its best plan. This ensures that
we micro-optimize and macro-optimize at the same time. For the best
macro-plan candidate, we focus on it, so we may as well use those
iterations to learn something about the best micro-plan for it. And we
don't waste time exploring the best micro-plan for macro-plans that are
no good.

For my "list all Coldplay tracks" test query, this new optimizer does
not always produce the same plan, but the macro plan is sound, and the
micro plans are really close. I ran it twice, and what is interesting,
is that it switches the order of two scans, and at the same time also
the index. I'd need to run it more often to see if there is a pattern
to it, but I'll include the results so far here for future reference.

Query:

    where
      t track.number number
      t track.title title
      t track.artist artist
      t track.album a
      a album.artist ar
      ar artist.name "Coldplay"
      a album.year year
    select
      number, title, artist, year

Optimized plan, first attempt:

    [6/7] [60 open] Candidate 309.294µs (n=5000)
    Micro 301.7 µs .. 309.294µs (n=1650)
    Micro 301.7 µs .. 310.479µs (n=1245)
    Micro 301.7 µs .. 319.369µs (n=326)
    Micro 301.7 µs .. 320.365µs (n=294)
    Micro 301.7 µs .. 311.981µs (n=918)
    Micro 301.7 µs .. 320.87µs (n=280)
    Micro 301.7 µs .. 332.044µs (n=120)
    Micro 301.7 µs .. 327.055µs (n=167)
    304.9 µs .. 320.584µs (n=419)
    304.9 µs .. 319.734µs (n=466)
    304.9 µs .. 320.847µs (n=406)

    Optimization complete.
    Duration: 321.645µs
    let :7 = Value(13835058055282163753)
    for ar:5 in avet (value=:7)
    for a:4 in avet (value=ar:5)
    for year:6 in eavt (entity=a:4)
    for t:1 in avet (value=a:4)
    for number:0 in eavt (entity=t:1)
    for artist:3 in eavt (entity=t:1)
    for title:2 in eavt (entity=t:1)
    yield

Optimized plan, second attempt:

    [6/7] [60 open] Candidate 308.769µs (n=5000)
    Micro 300.9 µs .. 309.707µs (n=1248)
    Micro 300.9 µs .. 310.962µs (n=963)
    Micro 300.9 µs .. 308.769µs (n=1556)
    Micro 300.9 µs .. 319.26µs (n=304)
    Micro 301.0 µs .. 331.329µs (n=119)
    Micro 301.0 µs .. 317.932µs (n=351)
    Micro 301.0 µs .. 319.782µs (n=289)
    Micro 301.0 µs .. 325.994µs (n=170)
    304.4 µs .. 319.291µs (n=460)
    304.4 µs .. 318.95µs (n=481)
    304.4 µs .. 320.066µs (n=418)

    Optimization complete.
    Duration: 322.564µs
    let :7 = Value(13835058055282163753)
    for ar:5 in avet (value=:7)
    for a:4 in avet (value=ar:5)
    for year:6 in eavt (entity=a:4)
    for t:1 in avet (value=a:4)
    for artist:3 in aevt (entity=t:1)
    for number:0 in aevt (entity=t:1)
    for title:2 in eavt (entity=t:1)
    yield